+++ GetPackages.BAT
@@ -1,0 +1,2 @@
+@Echo ON
+nuget install .\packages.config -OutputDirectory .\packages -excludeVersion
