--- PDT.EssentialsPluginTemplate.EPI/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("EssentialsPluginTemplateEpi")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("EssentialsPluginTemplateEpi")]
-[assembly: AssemblyCopyright("Copyright ©  2022")]
-[assembly: AssemblyVersion("1.0.0.*")]
-
