--- PDT.EssentialsPluginTemplate.EPI/SIMPLSharpLogs/(2019-08-30 15-01-01).log
@@ -1,63 +1,0 @@
-8/30/2019 3:01:01 PM, Info: Initializing SIMPLSharp Services...
-8/30/2019 3:01:01 PM, Info: ProjectInfo successfully initialized.
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:01:02 PM, Info: Saving project information...
-8/30/2019 3:04:48 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\essentials-builds\PepperDashEssentials.dll...
-8/30/2019 3:05:10 PM, Info: Saving project information...
-8/30/2019 3:05:10 PM, Info: Saving project information...
-8/30/2019 3:05:10 PM, Info: Saving project information...
-8/30/2019 3:05:39 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\essentials-builds\PepperDash_Core.dll...
-8/30/2019 3:05:39 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\essentials-builds\PepperDash_Essentials_Core.dll...
-8/30/2019 3:05:40 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\essentials-builds\PepperDash_Essentials_DM.dll...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:24 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:06:36 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:08:54 PM, Info: Saving project information...
-8/30/2019 3:09:39 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\essentials-builds\PepperDash_Essentials_Core.dll...
-8/30/2019 3:10:10 PM, Info: Saving project information...
-8/30/2019 3:10:10 PM, Info: Saving project information...
-8/30/2019 3:10:10 PM, Info: Saving project information...
-8/30/2019 3:15:10 PM, Info: Saving project information...
-8/30/2019 3:15:10 PM, Info: Saving project information...
-8/30/2019 3:15:10 PM, Info: Saving project information...
-8/30/2019 3:20:10 PM, Info: Saving project information...
-8/30/2019 3:20:10 PM, Info: Saving project information...
-8/30/2019 3:20:10 PM, Info: Saving project information...
-8/30/2019 3:25:10 PM, Info: Saving project information...
-8/30/2019 3:25:10 PM, Info: Saving project information...
-8/30/2019 3:25:10 PM, Info: Saving project information...
-8/30/2019 3:30:10 PM, Info: Saving project information...
-8/30/2019 3:30:10 PM, Info: Saving project information...
-8/30/2019 3:30:10 PM, Info: Saving project information...
-8/30/2019 3:35:10 PM, Info: Saving project information...
-8/30/2019 3:35:10 PM, Info: Saving project information...
-8/30/2019 3:35:10 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:35:38 PM, Info: Saving project information...
-8/30/2019 3:36:52 PM, Info: Validating assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\NvxUrlRoutingClientEpi\bin\Debug\NvxUrlRoutingClientEpi.dll...
-8/30/2019 3:36:52 PM, Info: Verifying assembly C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\NvxUrlRoutingClientEpi\bin\Debug\NvxUrlRoutingClientEpi.dll
-8/30/2019 3:36:52 PM, Info: Creating Archive C:\Users\JTA\Documents\Stash Folder\KPMG\SLN103-AVNoc\SLN103-AVNoc.Plugins\NvxUrlRoutingClient.Epi\NvxUrlRoutingClientEpi\bin\Debug\NvxUrlRoutingClientEpi.cplz...
-8/30/2019 3:36:53 PM, Info: Saving project information...
-9/2/2019 8:34:46 AM, Info: Terminating SIMPLSharp Services
