+++ src/PDT.EssentialsPluginTemplate.nuspec
@@ -2,6 +2,7 @@
 <package >
   <metadata>
     <id>PDT.EssentialsPluginTemplate</id>
+    <readme>docs\README.md</readme>
     <version>2.0.2</version>
     <title>Essentials Plugin Template</title>
     <authors>PepperDash Technologies</authors>
@@ -11,11 +12,12 @@
     <projectUrl>https://github.com/PepperDash/EssentialsPluginTemplate</projectUrl>
     <copyright>Copyright 2022</copyright>
     <description>This software iserves as a template for generating new Essentials-Compliant plugins</description>
-    <tags>crestron 3series 4series samsung</tags>
+    <tags>crestron 3series 4series essentials pepperdash pepperdashessentials epi essentials plugin</tags>
     <repository type="git" url="https://github.com/PepperDash/EssentialsPluginTemplate"/>
   </metadata>
     <files>
       <file src="**" target="lib\net35"/>
       <file src="**" target="lib\net47"/>
+      <file src="README.md" target="docs\"/>
     </files>
 </package>
