+++ src/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("EssentialsPluginTemplateEpi")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("EssentialsPluginTemplateEpi")]
+[assembly: AssemblyCopyright("Copyright ©  2022")]
+[assembly: AssemblyVersion("1.0.0.*")]
+
